+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 5.2.3 (unreleased)
 ------------------
 
+- Fix data corruption due to race between load and external invalidations.
+  See `issue 155 <https://github.com/zopefoundation/ZEO/issues/155>`_.
 
 5.2.2 (2020-08-11)
 ------------------