+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Fix data corruption due to race between load and external invalidations.
   See `issue 155 <https://github.com/zopefoundation/ZEO/issues/155>`_.
 
+- Improve log message when client cache is out of sync with server.
+  See `issue 142 <https://github.com/zopefoundation/ZEO/issues/142>`_.
+
 5.2.2 (2020-08-11)
 ------------------